refactor: change inheritance to delegation
